requirement.txt
streamlit==1.38.0
python-dotenv==1.0.1
langchain-google-genai==1.0.10